+++ dashboard/heart_rate_calculator.py
@@ -1,4 +1,4 @@
-def calculate_max_heart_rate(age: int) -> int:
+def calculate_max_heart_rate(age: float) -> float:
     """Calculates maximum heart rate given age. 211, 208, 0.64 and 0.7 taken from formulas found in published papers"""
     if age <= 40:
         return 211 - (0.64 * age)
@@ -6,7 +6,7 @@ def calculate_max_heart_rate(age: int) -> int:
         return 208 - (0.7 * age)
 
 
-def heart_rate_low(current_heart_rate: int, age: int) -> bool:
+def heart_rate_low(current_heart_rate: float, age: float) -> bool:
     """Returns true if current heart rate is too low"""
     max_heart_rate = calculate_max_heart_rate(age)
     lower_limit = 0.5 * max_heart_rate
